config file relocated to core dir, overall adjustments
