+++ .env.example
@@ -1,3 +1,4 @@
+DEBUG=
 TELEGRAM_TOKEN=
 NOTION_TOKEN=
 NOTION_DATABASE_ID=
