+++ README.md
@@ -55,10 +55,14 @@ $ alembic upgrade head
 ```
 
 ## Конфигурация проекта
-Валидирование всех переменных окружения проекта осуществляется через файл ```config.py```, находящемся в директории /app.
+Валидирование всех переменных окружения проекта осуществляется через файл ```config.py```, находящемся в директории /core.
+Если какие-то перменные окружения не были вами заданы в .env файле - в проекте будут использованы значения, заданные в файле ```config.py```.
+
+Для запука в бота в режиме разработки запускаем бот через пулинг.
+DOMAIN_ADDRESS понадобится только на стадиях test и production.
 
 Перед финальным деплоем проекта не забудьте выключить режим дебага:
-```DEBUG_MODE = False```
+```DEBUG=False```
 
 ### Важно
 Если вы добавляете или удаляете какую-то переменную окружения - обязательно отобразите эти изменения в файле config.py
