+++ app/main.py
@@ -4,10 +4,7 @@ from telegram import Update
 
 import bot
 from .api import router
-from .config import config
 
-
-settings = config['test']()
 
 def create_app() -> FastAPI:
     application = FastAPI()
@@ -19,13 +16,8 @@ app = create_app()
 
 dispatcher = bot.init()
 
+
 @app.post('/{token}/telegramWebhook')
 def webhook(data: dict):
     update = Update.de_json(data, dispatcher.bot)
     dispatcher.process_update(update)
-
-
-# тестоый эндпойнт для просмотра настроек
-@app.get('/settings')
-def get_settings():
-    return settings
